% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{add_copy_number_to_mutations}
\alias{add_copy_number_to_mutations}
\title{Add copy number data to mutations for INCOMMON analysis}
\usage{
add_copy_number_to_mutations(incommon_data)
}
\arguments{
\item{incommon_data}{Object from prepare_incommon_data()}
}
\value{
incommon_ready object with copy number added to mutations
}
\description{
Add copy number data to mutations for INCOMMON analysis
}
