% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bind_rows_safe}
\alias{bind_rows_safe}
\title{Combine data frames with different columns}
\usage{
bind_rows_safe(df_list)
}
\arguments{
\item{df_list}{List of data frames to combine}
}
\value{
Combined data frame with all columns
}
\description{
Combine data frames with different columns
}
\examples{
\dontrun{
df1 <- data.frame(a = 1:3, b = 4:6)
df2 <- data.frame(a = 7:9, c = 10:12)
bind_rows_safe(list(df1, df2))
}
}
