% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridge.R
\name{chordr_to_incommon}
\alias{chordr_to_incommon}
\title{Convert chordR output to INCOMMON input format}
\usage{
chordr_to_incommon(incommon_data, cohort)
}
\arguments{
\item{incommon_data}{Output from chordR::prepare_incommon_data() + add_copy_number_to_mutations()}

\item{cohort}{Output from chordR::extract_by_cancer_type()}
}
\value{
List with genomic_data and clinical_data ready for INCOMMON::init()
}
\description{
Convert chordR output to INCOMMON input format
}
