% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{create_timeline_object}
\alias{create_timeline_object}
\title{Create timeline object from extracted data}
\usage{
create_timeline_object(timeline_data, source_mae)
}
\arguments{
\item{timeline_data}{List of timeline data frames}

\item{source_mae}{Original MultiAssayExperiment object}
}
\value{
MSKTimeline object
}
\description{
Create timeline object from extracted data
}
\keyword{internal}
