% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_data}
\alias{download_data}
\title{Download data from cBioPortal}
\usage{
download_data(study = "msk_chord_2024")
}
\arguments{
\item{study}{The identifier of the study on cBioPortal}
}
\value{
A MultiAssayExperiment object
}
\description{
Download data from cBioPortal
}
