% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{estimate_eta}
\alias{estimate_eta}
\title{Estimate eta parameter for INCOMMON}
\usage{
estimate_eta(mutations)
}
\arguments{
\item{mutations}{Data frame with mutation data including k_total, VAF, t_depth, PURITY}
}
\value{
Numeric value of estimated eta
}
\description{
Estimates the read rate per chromosome copy (eta) from diploid regions
}
