% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_all_timelines}
\alias{extract_all_timelines}
\title{Extract all timeline types}
\usage{
extract_all_timelines(metadata_obj, timeline_configs)
}
\arguments{
\item{metadata_obj}{Metadata from MultiAssayExperiment}

\item{timeline_configs}{List of timeline configurations}
}
\value{
Named list of timeline data frames
}
\description{
Extract all timeline types
}
\keyword{internal}
