% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_single_timeline}
\alias{extract_single_timeline}
\title{Extract single timeline type}
\usage{
extract_single_timeline(metadata_obj, config)
}
\arguments{
\item{metadata_obj}{Metadata from MultiAssayExperiment}

\item{config}{Timeline configuration list}
}
\value{
Data frame with timeline data or NULL
}
\description{
Extract single timeline type
}
\keyword{internal}
