% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_timeline}
\alias{filter_timeline}
\title{Filter MSKTimeline object by multiple criteria}
\usage{
filter_timeline(
  timeline_obj,
  patients = NULL,
  date_range = NULL,
  event_types = NULL
)
}
\arguments{
\item{timeline_obj}{MSKTimeline object}

\item{patients}{Character vector of patient IDs to include}

\item{date_range}{Numeric vector of length 2: c(min_days, max_days)}

\item{event_types}{Character vector of event types to include}
}
\value{
Filtered MSKTimeline object
}
\description{
Filter MSKTimeline object by multiple criteria
}
\examples{
\dontrun{
filtered <- filter_timeline(
  timeline_obj,
  patients = c("P-0001", "P-0002")
)
}
}
