% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_subset.R
\name{get_clinical_events}
\alias{get_clinical_events}
\title{Get Clinical Events for Timeline Plotting}
\usage{
get_clinical_events(msk_data, patient_ids)
}
\arguments{
\item{msk_data}{A MultiAssayExperiment object from download_data()}

\item{patient_ids}{Character vector of patient IDs}
}
\value{
A data.frame with clinical events
}
\description{
Get Clinical Events for Timeline Plotting
}
\keyword{internal}
