% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{get_default_timeline_configs}
\alias{get_default_timeline_configs}
\title{Get default timeline configurations}
\usage{
get_default_timeline_configs()
}
\value{
List of timeline configurations
}
\description{
Get default timeline configurations
}
\keyword{internal}
