% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{get_mutations_by_gene}
\alias{get_mutations_by_gene}
\title{Get mutation data for specific genes from cohort}
\usage{
get_mutations_by_gene(cohort, genes)
}
\arguments{
\item{cohort}{msk_cancer_cohort object}

\item{genes}{Character vector of gene symbols}
}
\value{
Data frame with mutations in specified genes
}
\description{
Get mutation data for specific genes from cohort
}
