% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_patient_ids}
\alias{get_patient_ids}
\title{Get unique patient IDs from timeline object}
\usage{
get_patient_ids(timeline_obj)
}
\arguments{
\item{timeline_obj}{MSKTimeline object}
}
\value{
Character vector of unique patient IDs
}
\description{
Get unique patient IDs from timeline object
}
\examples{
\dontrun{
# Create example timeline
tl <- new("MSKTimeline", 
          data = list(labs = data.frame(PATIENT_ID = c("P1", "P2"))),
          n_patients = 2L)
get_patient_ids(tl)
}
}
