% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interactive.R
\name{launch_timeline_explorer}
\alias{launch_timeline_explorer}
\title{Launch Shiny Timeline Explorer App}
\usage{
launch_timeline_explorer(msk_data, port = NULL, launch.browser = TRUE)
}
\arguments{
\item{msk_data}{A MultiAssayExperiment object from download_data()}

\item{port}{Integer specifying the port to run the app on (default: random)}

\item{launch.browser}{Logical indicating whether to launch the app in browser}
}
\value{
Shiny app object
}
\description{
Launch Shiny Timeline Explorer App
}
\examples{
\dontrun{
msk_data <- download_data()
launch_timeline_explorer(msk_data)
}
}
