% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{MSKTimeline-class}
\alias{MSKTimeline-class}
\title{MSK-CHORD Timeline Data Class}
\description{
An S4 class to represent timeline data from MSK-CHORD dataset
}
\section{Slots}{

\describe{
\item{\code{data}}{Named list of timeline data frames}

\item{\code{metadata}}{List of metadata information}

\item{\code{cancer_types}}{Character vector of included cancer types}

\item{\code{n_patients}}{Integer number of unique patients}
}}

