% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plot_timeline}
\alias{plot_timeline}
\title{Create timeline visualization}
\usage{
plot_timeline(timeline_obj, patients = NULL)
}
\arguments{
\item{timeline_obj}{MSKTimeline object}

\item{patients}{Character vector of patient IDs to plot (max 20)}
}
\value{
ggplot2 object
}
\description{
Create timeline visualization
}
\examples{
\dontrun{
plot_timeline(timeline_data)
}
}
