% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridge.R
\name{print.chordr_incommon_bridge}
\alias{print.chordr_incommon_bridge}
\title{Print method for chordr_incommon_bridge}
\usage{
\method{print}{chordr_incommon_bridge}(x, ...)
}
\description{
Print method for chordr_incommon_bridge
}
