% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{show,MSKTimeline-method}
\alias{show,MSKTimeline-method}
\title{Show method for MSKTimeline}
\usage{
\S4method{show}{MSKTimeline}(object)
}
\arguments{
\item{object}{An MSKTimeline object}
}
\description{
Show method for MSKTimeline
}
