% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{standardize_timeline_columns}
\alias{standardize_timeline_columns}
\title{Standardize timeline column names}
\usage{
standardize_timeline_columns(data)
}
\arguments{
\item{data}{Data frame from timeline}
}
\value{
Data frame with standardized columns
}
\description{
Standardize timeline column names
}
\keyword{internal}
