% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_subset.R
\name{subset_by_cancer_type}
\alias{subset_by_cancer_type}
\title{Subset MSK-CHORD Data by Cancer Type}
\usage{
subset_by_cancer_type(msk_data, cancer_type)
}
\arguments{
\item{msk_data}{A MultiAssayExperiment object from download_data()}

\item{cancer_type}{Character string specifying the cancer type to subset.
Available types: "Breast Cancer", "Colorectal Cancer", "Non-Small Cell Lung Cancer",
"Pancreatic Cancer", "Prostate Cancer"}
}
\value{
A MultiAssayExperiment object subsetted to the specified cancer type
}
\description{
Subset MSK-CHORD Data by Cancer Type
}
\examples{
\dontrun{
msk_data <- download_data()
prostate_data <- subset_by_cancer_type(msk_data, "Prostate Cancer")
lung_data <- subset_by_cancer_type(msk_data, "Non-Small Cell Lung Cancer")
}
}
