% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{summary,MSKTimeline-method}
\alias{summary,MSKTimeline-method}
\title{Summary method for MSKTimeline}
\usage{
\S4method{summary}{MSKTimeline}(object)
}
\arguments{
\item{object}{An MSKTimeline object}
}
\description{
Summary method for MSKTimeline
}
