% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interactive.R
\name{timeline_plot_Server}
\alias{timeline_plot_Server}
\title{Interactive Timeline Plot Shiny Module Server}
\usage{
timeline_plot_Server(id, msk_data)
}
\arguments{
\item{id}{Character string, namespace identifier for the module}

\item{msk_data}{Reactive expression containing the MSK-CHORD data}
}
\value{
Server logic for the timeline plot module
}
\description{
Interactive Timeline Plot Shiny Module Server
}
\examples{
\dontrun{
# In your Shiny server
timeline_plot_Server("timeline_module", reactive(msk_data))
}
}
