% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interactive.R
\name{timeline_plot_UI}
\alias{timeline_plot_UI}
\title{Interactive Timeline Plot Shiny Module UI}
\usage{
timeline_plot_UI(id)
}
\arguments{
\item{id}{Character string, namespace identifier for the module}
}
\value{
Shiny UI elements for the timeline plot module
}
\description{
Interactive Timeline Plot Shiny Module UI
}
\examples{
\dontrun{
# In your Shiny UI
timeline_plot_UI("timeline_module")
}
}
