% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{validate_msk_chord_data}
\alias{validate_msk_chord_data}
\title{Validate MSK-CHORD data structure}
\usage{
validate_msk_chord_data(data)
}
\arguments{
\item{data}{MultiAssayExperiment object}
}
\value{
Invisible validation results
}
\description{
Validate MSK-CHORD data structure
}
\keyword{internal}
