% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridge.R
\name{verify_incommon_compatibility}
\alias{verify_incommon_compatibility}
\title{Verify chordR data compatibility with INCOMMON}
\usage{
verify_incommon_compatibility(incommon_data)
}
\arguments{
\item{incommon_data}{Output from prepare_incommon_data()}
}
\value{
Logical indicating if data is compatible
}
\description{
Verify chordR data compatibility with INCOMMON
}
